// atf-application-ui/src/app/model/scenario.ts
/*
 * Copyright 2018 BSC Msc, LLC 
 *
 * This file is part of the ATF project 
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); 
 * you may not use this file except in compliance with the License. 
 * You may obtain a copy of the License at 
 *
 * http://www.apache.org/licenses/LICENSE-2.0 
 *
 * Unless required by applicable law or agreed to in writing, software 
 * distributed under the License is distributed on an "AS IS" BASIS, 
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
 * See the License for the specific language governing permissions and 
 * limitations under the License.
 */

export class Scenario {
  code: string;
  projectCode: string;
  projectName: string;
  name: string;
  scenarioGroup: string;

  beforeScenarioIgnore: boolean;
  afterScenarioIgnore: boolean;

  failed: boolean;
  hasResults: boolean;

  _selected = false;
}